+++ ReleaseNotes.md
@@ -10,6 +10,7 @@
 - Removed named modules APIs
 - Removed IPipeline.ProcessDocumentsOnce and related plumbing
 - Split Execute module into Execute, ExecuteDocument, and ExecuteContext
+- Pipelines no longer start with an empty initial document (this was important before the execution context provided access to settings but it no longer needed)
 
 # 2.2.4